Suppress errorbells when they are called without count.
Since there is nothing to clear when they are called without count, <Esc> may cause errorbell in normal mode. Thus use <C-u> instead. In visual mode, <Esc> is sure to be used for the transition from visual mode to normal mode, and it is essential to store the correct cursor position for the "cursor" option. keep it.
This problem is reported in issue #9.