Hide cursor when waiting for an input in visual mode.
When user calls the operators in visual mode, they would wait for an user input in normal mode. At the time cursor has not been hidden, since it is not in operator-pending mode. Probably when pushing g@ key sequence to invoke operator function in visual mode, it immediately leads to normal mode.