Limit the number of highlighting region to one explicitly.
Originally the highlight object was designed to be able to highlight and scheduled-quench multiple places. However it is not helpful for practical use. So limit the number of highlighting region to one explicitly to avoid trivial weird behaviour. For example, ':normal saiw(saiw(' would show highlight differently before and after this commit.