Add backtick (`) to default_recipes
I think the backtick (`) should be on the default recipes, since they are frequently used in languages like javascript or bash. I thought I would offer the suggestion with this PR, but feel free to reject it if doesn't make sense to you

Thanks for `sandwich.vim` it's really a great plugin.