Elect textobjects by representative length
Textobj-auto has elected text by wrapped text excluding surroundings. However, this method works unintentionally, for example, reported at #127. This commit changes the election algorithm to use representative lengths; it is the length of wrapped text excluding surroundings if the cursor inside or including surroundings if the cursor is on either one of the surroundings.